% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARscore.R
\name{ARscore_algorithm}
\alias{ARscore_algorithm}
\title{Calculate Aggregate Reactivity Scores}
\usage{
ARscore_algorithm(
  hfc,
  fc,
  set_max_iterations = 10,
  set_p_cutoff = -log10(0.001),
  set_score_cutoff = 0.72,
  required_number_of_peptides = 50
)
}
\arguments{
\item{hfc}{a dataframe containing hits fold change data from a PhIP-Seq
experiment. Used to exclude peptides that are reactive in beads only controls}

\item{fc}{a dataframe containing fold change data from a PhIP-Seq experiment.
Used to generate intermediate reactivity metrics, null distributions, and 
ARscores}

\item{set_max_iterations}{integer limiting the number of itererations to run
calc_scores. defaults to 10}

\item{set_p_cutoff}{numeric -log10 p value cutoff to determine significant
targets of antibody reactivity. defaults to 4}

\item{set_score_cutoff}{numeric aggregate reactivity score cutoff to 
determine signifiant targets of antibody reactivity. defaults to .72 the 
empircally determined threshold for VRC VirScan data. Recommended to adjust}

\item{required_number_of_peptides}{integer that represents the minimum number
of peptides attributed to an antigen/pathogen required to calculate an 
ARscore}
}
\value{
a dataframe containing the aggregate reactivity scores for a PhIP run
}
\description{
Calculate Aggregate Reactivity Scores
}
