% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_toxscan_fc}
\alias{example_toxscan_fc}
\title{Example ToxScan fold change antibody profile}
\format{
## `example_virscan_fc`
A data frame
}
\source{
Larman Lab
}
\usage{
example_toxscan_fc
}
\description{
Data generated from one serum sample by the Larman Lab
}
\keyword{datasets}
