% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_virscan_fc}
\alias{example_virscan_fc}
\title{Example VirScan fold change antibody profile}
\format{
## `example_virscan_fc`
A data frame
}
\source{
Larman Lab
}
\usage{
example_virscan_fc
}
\description{
Data generated from one serum sample by the Larman Lab
}
\keyword{datasets}
