% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARscore.R
\name{\%nin\%}
\alias{\%nin\%}
\title{Left Right function to evaluate NOT left %in% right}
\usage{
left \%nin\% right
}
\arguments{
\item{left}{vector}

\item{right}{vector}
}
\value{
binary vector
}
\description{
Left Right function to evaluate NOT left %in% right
}
