% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARscore.R
\name{\%+\%}
\alias{\%+\%}
\title{Left Right function to paste two strings}
\usage{
left \%+\% right
}
\arguments{
\item{left}{string}

\item{right}{string}
}
\value{
concatenated string
}
\description{
Left Right function to paste two strings
}
