% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARscore.R
\name{iterative_scores}
\alias{iterative_scores}
\title{Shell function to iterate calculations of aggregate reactivity scores}
\usage{
iterative_scores(
  norm_log_1,
  all_peptide_fcs_1,
  max_iterations = 10,
  p_cutoff = -log10(1e-04),
  score_cutoff = 2,
  exclusion_method_1 = "genus"
)
}
\arguments{
\item{norm_log_1}{dataframe containing intermediate reactivity metrics}

\item{all_peptide_fcs_1}{dataframe containing all peptide fold changes 
for an individual antibody profile}

\item{max_iterations}{integer limiting the number of itererations to run
calc_scores. defaults to 10}

\item{p_cutoff}{numeric -log10 p value cutoff to determine significant
targets of antibody reactivity. defaults to 4}

\item{score_cutoff}{numeric aggregate reactivity score cutoff to determine
signifiant targets of antibody reactivity. defaults to 2 the empircally 
determined threshold for VRC VirScan data. Recommended to adjust}

\item{exclusion_method_1}{string describing method for excluding likely reactive
peptides. Defaults to by genus unless exclusion_method = 'group' or 'species"}
}
\value{
returns aggregate reactivity scores as well as a list denoting the
significant targets of reactivity detected during each iteration
}
\description{
v 1.0 update - two internal hit criteria are now used
}
